Working towards getting properties controllers implemented
Device and System properties are starting to be implemented.

Accounts and Music Services are being improved upon.
Account provisioning is not implemented yet, but will be.

Searching for tracks that can be played is soon to be implemented.

SOAP stuff is being abstracted into the SoapMessage class which should be far easier to use.

There's an LEDState PSVM that will let you toggle the LED state of a (randomly selected) Sonos speaker in your network! More on this later.